Add typescript. Move ./lib to ./src
