+++ .gitignore
@@ -12,8 +12,8 @@ logs
 results
 
 npm-debug.log
-
 node_modules
+build
 
 .idea
 *.iml
