+++ package.json
@@ -106,6 +106,7 @@
     "nyc": "^15.1.0",
     "should": "^13.2.3",
     "strip-ansi": "^6.0.0",
+    "typescript": "^4.2.4",
     "yarn": "^1.22.17"
   },
   "files": [
