+++ src/bin/cli.js
