+++ src/merge-options.ts
@@ -1,12 +1,13 @@
-function mergeArrays(arr1, arr2) {
-  return [...new Set([...arr1 || [], ...arr2 || []])]
+/** Merges two arrays into one, removing duplicates. */
+function mergeArrays(arr1: any[], arr2: any[]) {
+  return Array.from(new Set([...arr1 || [], ...arr2 || []]))
 }
 
 /**
  * Shallow merge (specific or all) properties.
  * If some properties both are arrays, then merge them also.
  */
-function mergeOptions(options1, options2) {
+function mergeOptions(options1: any, options2: any) {
   options1 = options1 || {}
   options2 = options2 || {}
   const result = { ...options1, ...options2 }
@@ -18,4 +19,4 @@ function mergeOptions(options1, options2) {
   return result
 }
 
-module.exports = mergeOptions
+export default mergeOptions
