+++ src/package-managers/index.js
