+++ src/package-managers/npm.js
