+++ src/versionmanager.js
