+++ test/deep.test.js
@@ -6,7 +6,7 @@ const chai = require('chai')
 const chaiAsPromised = require('chai-as-promised')
 const ncu = require('../lib/')
 const spawn = require('spawn-please')
-const mergeOptions = require('../lib/merge-options')
+const mergeOptions = require('../merge-options')
 
 chai.should()
 chai.use(chaiAsPromised)
